requirements.txt
Pillow==8.1.0
PyQt5==5.15.2
PyQt5-sip==12.8.1
PyQt5-stubs==5.15.2.0
